+++ lib/GuestPages/guest_shower.dart
@@ -2,6 +2,8 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+
+import '../model/user_model.dart';
 /*Page that allows guests to sign up for showers*/
 bool _hasBeenPressed = false;
 class GuestShowerPage extends StatefulWidget {
@@ -53,12 +55,23 @@ class _ShowerState extends State<GuestShowerPage> {
                               child: ElevatedButton(
                                 style: ElevatedButton.styleFrom(primary:_hasBeenPressed ? Colors.orange : Colors.green ),
                                 child: _hasBeenPressed ? const Text('Leave Line', style: TextStyle(fontSize: 20),): const Text('Join Line', style: TextStyle(fontSize: 20),),
-                                onPressed: () {
+                                onPressed: () async {
                                   setState(() {
                                     _hasBeenPressed = !_hasBeenPressed;
                                   });
-                                  joinLine();},
-
+                                  FirebaseFirestore.instance.collection("showers")
+                                      .doc(
+                                      FirebaseAuth.instance.currentUser!.uid)
+                                      .get()
+                                      .then((value) {
+                                        if(_hasBeenPressed == true && value.exists == false){
+                                          joinLine();
+                                        }
+                                        else{
+                                          leaveLine();
+                                        }
+                                  });
+                                }
                               ),
                             ),
                           )
@@ -74,14 +87,30 @@ class _ShowerState extends State<GuestShowerPage> {
 }
 
 void joinLine() {
+
   CollectionReference shower = FirebaseFirestore.instance.collection('showers');
 
   var currentUser = FirebaseAuth.instance.currentUser;
 
-  String? email = "";
+  String? firstName = "";
+  String? lastName = "";
+  String? uid = "";
 
   if (currentUser != null) {
-    email = currentUser.email;
+    FirebaseFirestore.instance.collection("users")
+        .doc(FirebaseAuth.instance.currentUser!.uid)
+        .get()
+        .then((value) {
+      firstName = UserModel
+          .fromMap(value.data())
+          .firstName;
+      lastName = UserModel
+          .fromMap(value.data())
+          .lastName;
+      uid = UserModel
+          .fromMap(value.data())
+          .uid;
+    });
   }
 
   FirebaseFirestore.instance
@@ -94,7 +123,8 @@ void joinLine() {
         .docs[0]["index"]; //Everything above here in the method is to find the highest previous index
     index = index + 1;
     shower.add({ //add new name to shower line with an incremented index
-      'name': email,
+      'name': firstName! + " " +  lastName!,
+      'uid' : uid,
       'index': index
     });
   });
@@ -122,6 +152,21 @@ void findSpot()
       .then((QuerySnapshot querySnapshot) {
     var index = querySnapshot.docs[0]["index"];
     print("The user's index is " + index.toString());
+  });
+
+}
+
+void leaveLine() {
+
+  String? uid = FirebaseAuth.instance.currentUser!.uid;
+
+  var showersRef = FirebaseFirestore.instance.collection('showers');
+
+  FirebaseFirestore.instance.collection('showers')
+      .where("uid", isEqualTo: uid)
+      .get()
+      .then((QuerySnapshot querySnapshot) {
+    showersRef.doc(querySnapshot.docs[0].id).delete();
   });
 
 }
